permit quota is int64

+++ src/dbnode/network/server/tchannelthrift/node/service.go
@@ -1093,9 +1093,9 @@ func (i *fetchTaggedResultsIter) Close(err error) {
 	i.seriesBlocks.RecordValue(float64(i.totalSeriesBlocks))
 
 	for n := 0; n < i.batchesAcquired-1; n++ {
-		i.blockPermits.Release(i.blocksPerBatch)
+		i.blockPermits.Release(int64(i.blocksPerBatch))
 	}
-	i.blockPermits.Release(i.blocksPerBatch - i.blocksAvailable)
+	i.blockPermits.Release(int64(i.blocksPerBatch - i.blocksAvailable))
 }
 
 // IDResult is the FetchTagged result for a series ID.
