+++ src/dbnode/storage/limits/permits/types.go
@@ -53,5 +53,5 @@ type Permits interface {
 	// Release gives back one acquired permit from the specific permits instance. The user can pass an optional quota
 	// indicating how much of quota was used while holding the permit.
 	// Cannot release more permits than have been acquired.
-	Release(quota int)
+	Release(quota int64)
 }
